Created views, and transfered functions properly
