+++ profiles/views.py
@@ -1,3 +1,14 @@
 from django.shortcuts import render
+from .models import Profile
 
-# Create your views here.
+
+def profiles_index(request):
+    profiles_list = Profile.objects.all()
+    context = {"profiles_list": profiles_list}
+    return render(request, "profiles/index.html", context)
+
+
+def profile(request, username):
+    profile = Profile.objects.get(user__username=username)
+    context = {"profile": profile}
+    return render(request, "profiles/profile.html", context)
